feat: added new file for handling station boards
